+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - git plugin will now always set remote and branch for pull and push
 - semver2 plugin always trying to tag a prerelease during bump action (#39)
 - changelog version sorting issues during generate and release (#38)
+- template plugin will now copy file permissions (#37)
 ### Removed
 - version plugin: automatic creation of dev tags (#46)